make hook installation work with submodules
So long as a submodule also has hooks installed at hooks/<name of hook>
within that submodule's project tree, this script (even if installed
*just in the parent module*, or it should work in the submodule too) may
be used to install the submodule's hooks, for that submodule.

In order for this to work correctly, the pwd that `autohook.sh install`
is called from must be the parent- or sub-module you want to install the
hooks for.

This scheme keeps submodule and parent module hooks separated even if
the same install script is used for both.